implemented redis services

+++ admin-service/src/controller.ts
@@ -3,6 +3,7 @@ import TryCatch from "./TryCatch.js";
 import getBuffer from "./config/dataUri.js";
 import cloudinary from "cloudinary";
 import { sql } from "./config/db.js";
+import { redisClient } from "./index.js";
 
 interface AuthenticatedRequest extends Request {
   user?: {
@@ -49,13 +50,18 @@ export const addAlbum = TryCatch(async (req: AuthenticatedRequest, res) => {
         INSERT INTO albums (title, description, thumbnail) VALUES (${title}, ${description}, ${cloud.secure_url}) RETURNING * 
     `;
 
-    res.json({
-        message : "Album created",
-        album: result[0]
-    })
+  if (redisClient.isReady) {
+    await redisClient.del("albums");
+    console.log("Cache invalidated for albums");
+  }
+
+  res.json({
+    message: "Album created",
+    album: result[0],
+  });
 });
 
-export const addSong = TryCatch(async (req:AuthenticatedRequest, res) => {
+export const addSong = TryCatch(async (req: AuthenticatedRequest, res) => {
   if (req.user?.role !== "admin") {
     res.status(401).json({
       message: "You are not admin",
@@ -66,14 +72,14 @@ export const addSong = TryCatch(async (req:AuthenticatedRequest, res) => {
 
   const { title, description, album } = req.body;
 
-  const isAlbum = await sql`SELECT * FROM albums WHERE id = ${album}`
+  const isAlbum = await sql`SELECT * FROM albums WHERE id = ${album}`;
 
-  if (isAlbum.length === 0){
+  if (isAlbum.length === 0) {
     res.status(404).json({
-      message: "No album with this id"
-    })
+      message: "No album with this id",
+    });
 
-    return
+    return;
   }
 
   const file = req.file;
@@ -98,7 +104,7 @@ export const addSong = TryCatch(async (req:AuthenticatedRequest, res) => {
 
   const cloud = await cloudinary.v2.uploader.upload(fileBuffer.content, {
     folder: "songs",
-    resource_type: "video"
+    resource_type: "video",
   });
 
   const result = await sql`
@@ -106,13 +112,17 @@ export const addSong = TryCatch(async (req:AuthenticatedRequest, res) => {
     (${title}, ${description}, ${cloud.secure_url}, ${album})
   `;
 
+  if (redisClient.isReady) {
+    await redisClient.del("songs");
+    console.log("Cache invalidated for songs");
+  }
+
   res.json({
-    message: "Song Added !"
-  })
+    message: "Song Added !",
+  });
+});
 
-})
-
-export const addThumbnail = TryCatch(async (req:AuthenticatedRequest, res) => {
+export const addThumbnail = TryCatch(async (req: AuthenticatedRequest, res) => {
   if (req.user?.role !== "admin") {
     res.status(401).json({
       message: "You are not admin",
@@ -121,14 +131,14 @@ export const addThumbnail = TryCatch(async (req:AuthenticatedRequest, res) => {
     return;
   }
 
-  const song = await sql`SELECT * FROM songs WHERE id = ${req.params.id}`
+  const song = await sql`SELECT * FROM songs WHERE id = ${req.params.id}`;
 
-  if (song.length === 0){
+  if (song.length === 0) {
     res.status(404).json({
-      message: "No song with this id"
-    })
+      message: "No song with this id",
+    });
 
-    return
+    return;
   }
 
   const file = req.file;
@@ -151,48 +161,24 @@ export const addThumbnail = TryCatch(async (req:AuthenticatedRequest, res) => {
     return;
   }
 
-  const cloud = await cloudinary.v2.uploader.upload(fileBuffer.content)
+  const cloud = await cloudinary.v2.uploader.upload(fileBuffer.content);
   const result = await sql`
     UPDATE songs SET thumbnail = ${cloud.secure_url} WHERE id = ${req.params.id} RETURNING *
-  `
+  `;
+
+  if (redisClient.isReady) {
+    await redisClient.del("songs");
+    console.log("Cache invalidated for songs");
+  }
 
   res.json({
-    messsage : "Thumbnail added!",
-    song : result[0]
-  })
-})
+    messsage: "Thumbnail added!",
+    song: result[0],
+  });
+});
 
-export const deleteAlbum = TryCatch(async (req:AuthenticatedRequest, res) => {
+export const deleteAlbum = TryCatch(async (req: AuthenticatedRequest, res) => {
   if (req.user?.role !== "admin") {
-    res.status(401).json({
-      message: "You are not admin",
-    });
-
-    return;
-  }
-
-   const {id} = req.params
-
-  const isAlbum = await sql`SELECT * FROM albums WHERE id = ${id}`
-
-  if (isAlbum.length === 0){
-    res.status(404).json({
-      message: "No album with this id"
-    })
-
-    return
-  }
-
-  await sql`DELETE FROM songs WHERE album_id = ${id}`
-  await sql`DELETE FROM albums WHERE id = ${id}`
-
-  res.json({
-    messsage: "Album deleted successfully!"
-  })
-})
-
-export const deleteSong = TryCatch(async (req:AuthenticatedRequest, res) => {
-    if (req.user?.role !== "admin") {
     res.status(401).json({
       message: "You are not admin",
     });
@@ -202,19 +188,63 @@ export const deleteSong = TryCatch(async (req:AuthenticatedRequest, res) => {
 
   const { id } = req.params;
 
-  const isSong = await sql`SELECT * FROM songs WHERE id = ${id}`
+  const isAlbum = await sql`SELECT * FROM albums WHERE id = ${id}`;
 
-  if (isSong.length === 0){
+  if (isAlbum.length === 0) {
     res.status(404).json({
-      message: "No album with this id"
-    })
+      message: "No album with this id",
+    });
 
-    return
+    return;
   }
 
-  await sql`DELETE FROM songs WHERE id = ${id}`
+  await sql`DELETE FROM songs WHERE album_id = ${id}`;
+  await sql`DELETE FROM albums WHERE id = ${id}`;
+
+  if (redisClient.isReady) {
+    await redisClient.del("albums");
+    console.log("Cache invalidated for albums");
+  }
+
+  if (redisClient.isReady) {
+    await redisClient.del("songs");
+    console.log("Cache invalidated for songs");
+  }
 
   res.json({
-    message: "song deleted!"
-  })
-})
+    messsage: "Album deleted successfully!",
+  });
+});
+
+export const deleteSong = TryCatch(async (req: AuthenticatedRequest, res) => {
+  if (req.user?.role !== "admin") {
+    res.status(401).json({
+      message: "You are not admin",
+    });
+
+    return;
+  }
+
+  const { id } = req.params;
+
+  const isSong = await sql`SELECT * FROM songs WHERE id = ${id}`;
+
+  if (isSong.length === 0) {
+    res.status(404).json({
+      message: "No album with this id",
+    });
+
+    return;
+  }
+
+  await sql`DELETE FROM songs WHERE id = ${id}`;
+
+  if (redisClient.isReady) {
+    await redisClient.del("songs");
+    console.log("Cache invalidated for songs");
+  }
+
+  res.json({
+    message: "song deleted!",
+  });
+});
